Move Logging to API
