+++ src/main/java/com/freya02/botcommands/api/SettingsProvider.java
@@ -1,7 +1,6 @@
 package com.freya02.botcommands.api;
 
 import com.freya02.botcommands.api.application.GuildApplicationSettings;
-import com.freya02.botcommands.internal.DefaultMessages;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.User;
 import org.jetbrains.annotations.NotNull;
