+++ src/main/java/com/freya02/botcommands/api/pagination/BasicPagination.java
@@ -1,8 +1,8 @@
 package com.freya02.botcommands.api.pagination;
 
 import com.freya02.botcommands.api.BContext;
+import com.freya02.botcommands.api.Logging;
 import com.freya02.botcommands.api.components.ComponentManager;
-import com.freya02.botcommands.internal.Logging;
 import com.freya02.botcommands.internal.utils.Utils;
 import net.dv8tion.jda.api.MessageBuilder;
 import net.dv8tion.jda.api.entities.Message;
